+++ config/splat.us.dra.yaml
@@ -41,6 +41,8 @@ segments:
       - [0x8258, data]
       - [0x8900, data]
       - [0xCEB0, data]
+      - [0xCF18, .data, 6D59C]
+      - [0xCF2C, data]
       - [0xD670, data]
       - [0x2217C, data]
       - [0x227B0, data]
@@ -182,6 +184,7 @@ segments:
       - [0x62D70, c, 62D70]
       - [0x673C0, c, cd]
       - [0x692E8, c, 692E8]
+      - [0x6D59C, c, 6D59C]
       - [0x71830, c, 71830]
       - [0x72BB0, c, 72BB0]
       - [0x73AAC, c, 73AAC]
